backend/Usermanagement-with-Django-and-Rest-Framework-API-main2/requirements.txt
asgiref==3.7.2
dj-user-login-history==1.0.5
Django==5.0.1
django-admin==2.0.2
django-cors-headers==4.3.1
django-excel-response2==3.0.6
django-six==1.0.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
excel-base==1.0.4
isoweek==1.3.3
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
screen==1.0.1
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
tzdata==2023.4
tzlocal==5.2
xlwt==1.3.0